Login and Logout functionality

+++ app/javascript/components/Home/Home.jsx
@@ -3,38 +3,42 @@ import TitleBanner from "./TitleBanner";
 import Filter from "./Filter";
 import ToDoList from "./ToDoList/ToDoList"
 import AddToDo from "./AddToDo/AddToDo";
+import Login from "./Login";
+import Logout from "./logout";
 
 class Home extends Component {
     constructor(props) {
         super(props);
         this.state = {
             to_dos: [
-                {
-                    id: 1,
-                    title: 'Laundry',
-                    description: 'Wash the whites.',
-                    status: "Pending",
-                    active: false,
-                    editing: false,
-                },
-                {
-                    id: 2,
-                    title: 'Build App',
-                    description: 'Build a To-Do-List app using ruby on rails and react.',
-                    status: "Pending",
-                    active: false,
-                    editing: false,
-                },
-                {
-                    id: 3,
-                    title: 'Write Speech',
-                    description: 'Write speech for graduation',
-                    status: "Complete",
-                    active: false,
-                    editing: false,
-                },
+                // {
+                //     id: 1,
+                //     title: 'Laundry',
+                //     description: 'Wash the whites.',
+                //     status: "Pending",
+                //     active: false,
+                //     editing: false,
+                // },
+                // {
+                //     id: 2,
+                //     title: 'Build App',
+                //     description: 'Build a To-Do-List app using ruby on rails and react.',
+                //     status: "Pending",
+                //     active: false,
+                //     editing: false,
+                // },
+                // {
+                //     id: 3,
+                //     title: 'Write Speech',
+                //     description: 'Write speech for graduation',
+                //     status: "Complete",
+                //     active: false,
+                //     editing: false,
+                // },
             ],
-            filter: "All"
+            filter: "All",
+            loggedIn: false,
+            username: ''
         }
     }
 
@@ -109,28 +113,55 @@ class Home extends Component {
 
     }
 
+    login(item, event) {
+        event.preventDefault();
+        let u = item();
+        if(item !== '')
+            this.setState({
+                username: u,
+                loggedIn: true
+            })
+    }
+
+    logout(event) {
+        this.setState({
+            to_dos: [],
+            filter: "All",
+            loggedIn: false,
+            username: ''
+        })
+
+    }
+
     render() {
-        return (
-            <div>
-                <TitleBanner/>
-                <div className='container w-75'>
-                    <div className='row'>
-                        <div className='col-8'>
-                            <Filter changeFilter={this.changeFilter.bind(this)}/>
-                            <ToDoList toggleActive={this.toggleActive.bind(this)} to_dos={this.state.to_dos}
-                                      filter={this.state.filter} markComplete={this.markComplete.bind(this)}
-                                      markPending={this.markPending.bind(this)}
-                                      saveChanges={this.saveChanges.bind(this)}
-                                      deleteItem={this.deleteItem.bind(this)}
-                                      toggleEdit={this.toggleEdit.bind(this)}/>
-                        </div>
-                        <div className='col-4'>
-                            <AddToDo addItem={this.addItem.bind(this)} nextId={this.state.to_dos.length + 1}/>
+        if(!this.state.loggedIn)
+            return(
+                <Login login={this.login.bind(this)}/>
+            );
+        else {
+            return (
+                <div>
+                    <Logout logout={this.logout.bind(this)}/>
+                    <TitleBanner/>
+                    <div className='container w-75'>
+                        <div className='row'>
+                            <div className='col-8'>
+                                <Filter changeFilter={this.changeFilter.bind(this)}/>
+                                <ToDoList toggleActive={this.toggleActive.bind(this)} to_dos={this.state.to_dos}
+                                          filter={this.state.filter} markComplete={this.markComplete.bind(this)}
+                                          markPending={this.markPending.bind(this)}
+                                          saveChanges={this.saveChanges.bind(this)}
+                                          deleteItem={this.deleteItem.bind(this)}
+                                          toggleEdit={this.toggleEdit.bind(this)}/>
+                            </div>
+                            <div className='col-4'>
+                                <AddToDo addItem={this.addItem.bind(this)} nextId={this.state.to_dos.length + 1}/>
+                            </div>
                         </div>
                     </div>
                 </div>
-            </div>
-        )
+            )
+        }
     }
 }
